+++ app/assets/scripts/editor.js
@@ -389,6 +389,23 @@
 
                 c6State.goTo('editor');
             };
+
+            $scope.$watch(
+                function() { return self.model.data.service; },
+                function(service, prevService) {
+                    var data = self.model.data;
+
+                    if (service === 'dailymotion') {
+                        data.start = undefined;
+                        data.end = undefined;
+                    }
+
+                    if (prevService === 'dailymotion') {
+                        data.start = null;
+                        data.end = null;
+                    }
+                }
+            );
         }])
 
         .controller('NewCardController', ['$scope','c6State','c6StateParams','MiniReelService',
@@ -817,6 +834,8 @@
 
                             startScanTime = null;
                         }
+
+                        scope.video = undefined;
 
                         if (!video) { return; }